# docs/openapi.yaml
openapi: 3.1.0
info:
  title: theserver API
  version: "1"
  description: |
    API v1 of theserver, the local venue server of CYB3RGUN. Every request
    except GET /openapi.yaml carries an admin token as a bearer token
    (D-025); a package download also takes the token of an approved device
    (D-038). A missing or unknown token is answered with 401, a revoked one
    with 403. Timestamps are unix milliseconds in UTC (D-014); 0 means not
    set. Errors always have the shape of the Error schema.
servers:
  - url: /api/v1
security:
  - adminToken: []
paths:
  /devices:
    get:
      summary: List every device with its link state.
      operationId: listDevices
      responses:
        "200":
          description: The devices, ordered by id.
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                properties:
                  devices:
                    type: array
                    items: {$ref: "#/components/schemas/Device"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /devices/{id}:
    get:
      summary: Read one device.
      operationId: getDevice
      parameters: [{$ref: "#/components/parameters/DeviceID"}]
      responses:
        "200":
          description: The device.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Device"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /devices/{id}/approve:
    post:
      summary: Approve a device, so it may connect.
      operationId: approveDevice
      parameters: [{$ref: "#/components/parameters/DeviceID"}]
      responses:
        "200":
          description: The device after the change.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Device"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /devices/{id}/block:
    post:
      summary: Block a device and drop its live connection.
      description: The connection gets err unauthorized and is closed; new connections are refused with 401.
      operationId: blockDevice
      parameters: [{$ref: "#/components/parameters/DeviceID"}]
      responses:
        "200":
          description: The device after the change.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Device"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /devices/{id}/reset:
    post:
      summary: Start a new sequence epoch for a device (D-026).
      description: |
        The device begins at seq 1 again; the events of earlier epochs stay.
        Its live connection is closed with WebSocket status 1012; the device
        reconnects and learns the new epoch from the ep key of welcome.
      operationId: resetDevice
      parameters: [{$ref: "#/components/parameters/DeviceID"}]
      responses:
        "200":
          description: The device with its new epoch.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Device"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /devices/{id}/token:
    post:
      summary: Replace the token of a device.
      description: |
        The new token is returned this once and only its hash is stored. The
        old token stops working at once: a live connection gets err
        unauthorized and is closed, and new connections with it get 401.
      operationId: newDeviceToken
      parameters: [{$ref: "#/components/parameters/DeviceID"}]
      responses:
        "200":
          description: The new token.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/NewToken"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /devices/{id}/min_age:
    post:
      summary: Set the age a device is set for (D-039).
      description: |
        A scenario rated above this age cannot be assigned to a session the
        device is in. The age is one of 0, 6, 12, 16 and 18; a new device is
        set for 18. An age below the rating of the scenario a created or
        running session of the device plays is refused with age_rating.
      operationId: setMinAge
      parameters: [{$ref: "#/components/parameters/DeviceID"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/MinAge"}
      responses:
        "200":
          description: The device after the change.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Device"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/ContentConflict"}
  /devices/{id}/scenarios:
    get:
      summary: The scenario versions a device holds or held, and the ones its sessions play.
      operationId: deviceScenarios
      parameters: [{$ref: "#/components/parameters/DeviceID"}]
      responses:
        "200":
          description: Holdings by scenario, newest version first, and the assignments of created and running sessions.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DeviceScenarios"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /sessions:
    get:
      summary: List every session, oldest first.
      operationId: listSessions
      responses:
        "200":
          description: The sessions.
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items: {$ref: "#/components/schemas/Session"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      summary: Create a session in state created.
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/NewSession"}
      responses:
        "201":
          description: The new session.
          headers:
            Location:
              schema: {type: string}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
  /sessions/{id}/start:
    post:
      summary: Move a created session to running.
      operationId: startSession
      parameters: [{$ref: "#/components/parameters/SessionID"}]
      responses:
        "200":
          description: The session after the change.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /sessions/{id}/stop:
    post:
      summary: Move a running session to stopped.
      operationId: stopSession
      parameters: [{$ref: "#/components/parameters/SessionID"}]
      responses:
        "200":
          description: The session after the change.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /sessions/{id}/devices:
    post:
      summary: Put a device into a session.
      description: |
        Adding a device twice is not an error. A device set for an age below
        the rating of the scenario the session plays is refused with
        age_rating; otherwise the device gets content_available when it does
        not hold that version.
      operationId: addSessionDevice
      parameters: [{$ref: "#/components/parameters/SessionID"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SessionDevice"}
      responses:
        "200":
          description: The session after the change.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/ContentConflict"}
  /sessions/{id}/scenario:
    post:
      summary: Assign a published scenario version to a session and announce it (D-038, D-039).
      description: |
        The session has to be created and the version published. No device
        of the session may be set for an age below the rating of the
        scenario. Every device of the session that does not hold the version
        gets content_available at once; a device that is offline gets it
        after its first health report on its next connection.
      operationId: assignScenario
      parameters: [{$ref: "#/components/parameters/SessionID"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ScenarioAssignment"}
      responses:
        "200":
          description: The session and what every device of it was told.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SessionAssignment"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/ContentConflict"}
  /scenarios:
    get:
      summary: The catalogue, one entry per scenario (D-036).
      operationId: listScenarios
      responses:
        "200":
          description: The scenarios by id, each with its latest published version and its drafts.
          content:
            application/json:
              schema:
                type: object
                required: [scenarios]
                properties:
                  scenarios:
                    type: array
                    items: {$ref: "#/components/schemas/ScenarioSummary"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      summary: Upload a scenario package; it is validated and kept as a draft (D-037).
      description: |
        The body is multipart/form-data with the zip in the field package,
        at most content.max_upload_mb megabytes. The manifest names the id
        and the version. A version above the latest published one is stored
        as a draft, replacing a draft of the same version; the draft keeps
        its validation problems and cannot be published until an upload
        without problems replaces it. A package that cannot be read, whose
        id or version is unusable, or whose version is published already is
        not stored and answered with 422 and its problems.
      operationId: uploadScenario
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [package]
              properties:
                package: {type: string, format: binary}
      responses:
        "201":
          description: A new draft.
          headers:
            Location:
              schema: {type: string}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Upload"}
        "200":
          description: A draft of the same version was replaced.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Upload"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "413":
          description: The package is larger than the upload limit; code too_large.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "422":
          description: The package was not stored; code invalid_package, with error.problems.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
  /scenarios/{id}:
    get:
      summary: Every version of a scenario and the devices that hold them.
      operationId: getScenario
      parameters: [{$ref: "#/components/parameters/ScenarioID"}]
      responses:
        "200":
          description: The versions, newest first, with their problems, and the holdings.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ScenarioDetail"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /scenarios/{id}/{version}:
    delete:
      summary: Delete a draft and its package; a published version stays (D-037).
      operationId: deleteScenario
      parameters:
        - {$ref: "#/components/parameters/ScenarioID"}
        - {$ref: "#/components/parameters/Version"}
      responses:
        "204":
          description: The draft is gone.
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/ContentConflict"}
  /scenarios/{id}/{version}/publish:
    post:
      summary: Publish a draft; from then on it never changes (D-037).
      description: |
        A draft with problems is refused with has_problems, a published
        version with published, and a draft at or below the latest published
        version with version_taken.
      operationId: publishScenario
      parameters:
        - {$ref: "#/components/parameters/ScenarioID"}
        - {$ref: "#/components/parameters/Version"}
      responses:
        "200":
          description: The published version.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ScenarioVersion"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/ContentConflict"}
  /scenarios/{id}/{version}/package.zip:
    get:
      summary: Download a package exactly as it was uploaded (D-038).
      description: |
        Devices fetch published versions with their device token; admins
        fetch any version. Range requests are answered with 206. The header
        X-Manifest-SHA256 carries the manifest hash, which content_available
        announces.
      operationId: downloadPackage
      security:
        - adminToken: []
        - deviceToken: []
      parameters:
        - {$ref: "#/components/parameters/ScenarioID"}
        - {$ref: "#/components/parameters/Version"}
        - name: Range
          in: header
          schema: {type: string, example: "bytes=1024-"}
      responses:
        "200":
          description: The whole package.
          headers:
            X-Manifest-SHA256:
              schema: {type: string}
          content:
            application/zip:
              schema: {type: string, format: binary}
        "206":
          description: The requested part of the package.
          headers:
            X-Manifest-SHA256:
              schema: {type: string}
            Content-Range:
              schema: {type: string}
          content:
            application/zip:
              schema: {type: string, format: binary}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403":
          description: The admin token was revoked, or the device is not approved.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "404": {$ref: "#/components/responses/NotFound"}
        "416":
          description: The range lies outside the package.
  /scenarios/{id}/{version}/cover.png:
    get:
      summary: The cover of a version, for the catalogue.
      operationId: scenarioCover
      parameters:
        - {$ref: "#/components/parameters/ScenarioID"}
        - {$ref: "#/components/parameters/Version"}
      responses:
        "200":
          description: cover.png of the package.
          content:
            image/png:
              schema: {type: string, format: binary}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /rankings:
    get:
      summary: Compute the ranking of a session or of everything (D-023).
      operationId: rankings
      parameters:
        - name: session
          in: query
          description: A session id; empty means all events, in a session or not.
          schema: {type: string}
      responses:
        "200":
          description: The ranking, by points, then hits, then controller id.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Ranking"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
  /events:
    get:
      summary: Read the journal, in the order the server received it.
      operationId: events
      parameters:
        - {name: device, in: query, schema: {type: string}}
        - {name: session, in: query, schema: {type: string}}
        - {name: kind, in: query, schema: {type: string}}
        - {name: controller, in: query, schema: {type: string}}
        - {name: from, in: query, description: ts_server from, inclusive., schema: {type: integer, minimum: 0}}
        - {name: to, in: query, description: ts_server to, exclusive., schema: {type: integer, minimum: 0}}
        - {name: limit, in: query, description: Default 100., schema: {type: integer, minimum: 0, maximum: 1000}}
        - {name: offset, in: query, schema: {type: integer, minimum: 0}}
      responses:
        "200":
          description: One page of events.
          content:
            application/json:
              schema:
                type: object
                required: [events, limit, offset]
                properties:
                  events:
                    type: array
                    items: {$ref: "#/components/schemas/Event"}
                  limit: {type: integer}
                  offset: {type: integer}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /online:
    get:
      summary: List the devices connected right now.
      operationId: online
      responses:
        "200":
          description: The connected devices, ordered by id.
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                properties:
                  devices:
                    type: array
                    items: {$ref: "#/components/schemas/Online"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
...
  /settings/reset:
    post:
      summary: Set settings back to their defaults, all or none.
      description: The settings are removed from the configuration file, otherwise as PUT /settings.
      operationId: resetSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [keys]
              additionalProperties: false
              properties:
                keys:
                  type: array
                  items: {type: string}
      responses:
        "200":
          description: The changes and what they need.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SettingsChanged"}
        "400": {$ref: "#/components/responses/SettingsRefused"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/FileAppeared"}
...
components:
  securitySchemes:
    adminToken:
      type: http
      scheme: bearer
      description: An admin token from `theserver admin token add`.
    deviceToken:
      type: http
      scheme: bearer
      description: The token of an approved device, for package downloads only.
  parameters:
    DeviceID:
      name: id
      in: path
      required: true
      schema: {type: string}
...
  responses:
    BadRequest:
      description: The request is malformed.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Unauthorized:
      description: No admin token, or an unknown one.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Forbidden:
      description: The admin token was revoked.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    NotFound:
      description: The device, session or endpoint does not exist.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
...
    ContentConflict:
      description: |
        The change conflicts with the scenario content: published, the
        version is published and stays; has_problems, the draft did not
        validate; version_taken, a newer version is published; not_published,
        a draft cannot be assigned; age_rating, the scenario is rated above
        the age a device is set for; bad_transition, the session is no longer
        created.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
...
  schemas:
...
    Device:
      type: object
      properties:
        id: {type: string}
        kind: {type: string, enum: [target, controller, bridge]}
        class: {type: string, enum: [esp, pi, pc]}
        name: {type: string}
        room: {type: string}
        zone: {type: string}
        status: {type: string, enum: [pending, approved, blocked]}
        firmware_version: {type: string}
        seq_epoch: {type: integer, minimum: 1}
        has_token: {type: boolean}
        first_seen: {type: integer}
        last_seen: {type: integer}
        created_at: {type: integer}
        updated_at: {type: integer}
        online: {type: boolean}
        connected_at: {type: integer, description: 0 while offline.}
        last_ack: {type: integer, description: 0 while offline.}
        min_age: {type: integer, enum: [0, 6, 12, 16, 18], description: The age the device is set for (D-039).}
    NewToken:
      type: object
      properties:
        device_id: {type: string}
        token: {type: string, description: Shown this once.}
        warning: {type: string}
...
    MinAge:
      type: object
      additionalProperties: false
      required: [min_age]
      properties:
        min_age: {type: integer, enum: [0, 6, 12, 16, 18]}
...